+++ src/main/java/xizi/Event.java
@@ -1,3 +1,5 @@
+package xizi;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
