+++ src/main/java/xizi/Parser.java
@@ -1,3 +1,7 @@
+package xizi;
+
+import xizi.command.*;
+
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
