+++ src/main/java/xizi/Storage.java
@@ -1,4 +1,4 @@
-// Do not use wildcard imports
+package xizi;// Do not use wildcard imports
 import java.io.IOException;
 import java.io.BufferedReader;
 import java.io.FileReader;
@@ -10,7 +10,6 @@ import java.nio.file.Paths;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 public class Storage {
