+++ src/main/java/xizi/XiziException.java
@@ -1,3 +1,5 @@
+package xizi;
+
 public class XiziException extends Exception{
     public XiziException(String message) {
         super(message);
