+++ src/main/java/xizi/command/Command.java
@@ -1,3 +1,10 @@
+package xizi.command;
+
+import xizi.Storage;
+import xizi.TaskList;
+import xizi.Ui;
+import xizi.XiziException;
+
 import java.io.IOException;
 
 public interface Command {
