+++ src/main/java/xizi/command/CommandType.java
@@ -1,3 +1,5 @@
+package xizi.command;
+
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 // Enums for command types
