+++ src/main/java/xizi/command/EventCommand.java
@@ -1,3 +1,7 @@
+package xizi.command;
+
+import xizi.*;
+
 import java.io.IOException;
 import java.util.regex.Matcher;
 import java.time.LocalDateTime;
